move historical func out of Scraper class
